Fix other deprecation warnings

+++ javalib/src/main/scala/java/lang/reflect/Array.scala
@@ -106,8 +106,8 @@ object Array {
     case array: Array[Char]  => array(index)
     case array: Array[Byte]  => array(index)
     case array: Array[Short] => array(index)
-    case array: Array[Int]   => array(index)
-    case array: Array[Long]  => array(index)
+    case array: Array[Int]   => array(index).toFloat
+    case array: Array[Long]  => array(index).toFloat
     case _ =>
       throw new IllegalArgumentException("argument type mismatch")
   }
@@ -118,7 +118,7 @@ object Array {
     case array: Array[Byte]   => array(index)
     case array: Array[Short]  => array(index)
     case array: Array[Int]    => array(index)
-    case array: Array[Long]   => array(index)
+    case array: Array[Long]   => array(index).toDouble
     case array: Array[Float]  => array(index)
     case _ =>
       throw new IllegalArgumentException("argument type mismatch")
@@ -182,7 +182,7 @@ object Array {
   def setInt(array: AnyRef, index: Int, value: Int): Unit = array match {
     case array: Array[Int]    => array(index) = value
     case array: Array[Long]   => array(index) = value
-    case array: Array[Float]  => array(index) = value
+    case array: Array[Float]  => array(index) = value.toFloat
     case array: Array[Double] => array(index) = value
     case _ =>
       throw new IllegalArgumentException("argument type mismatch")
@@ -190,8 +190,8 @@ object Array {
 
   def setLong(array: AnyRef, index: Int, value: Long): Unit = array match {
     case array: Array[Long]   => array(index) = value
-    case array: Array[Float]  => array(index) = value
-    case array: Array[Double] => array(index) = value
+    case array: Array[Float]  => array(index) = value.toFloat
+    case array: Array[Double] => array(index) = value.toDouble
     case _ =>
       throw new IllegalArgumentException("argument type mismatch")
   }
